backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
retina-face
tensorflow
tf-keras
insightface
filterpy
kagglehub
facenet-pytorch
torch>=2.0.0
torchvision
torchaudio
transformers
opencv-python
librosa
ffmpeg-python
timm
onnxruntime
numpy
pillow
grad-cam
sqlalchemy
alembic
psycopg2-binary
python-dotenv
scikit-learn
matplotlib
einops
pydub
static-ffmpeg
